+++ src/main/java/de/hhbk/jahresprojekt/model/Document.java
@@ -3,20 +3,20 @@ package de.hhbk.jahresprojekt.model;
 import java.util.List;
 
 public class Document {
-    private int id;
+    private Long id;
     private String fileName;
     private List<File> versionList;
 
     public Document() {
     }
 
-    public Document(int id, String fileName, List<File> versionList) {
+    public Document(Long id, String fileName, List<File> versionList) {
         this.id = id;
         this.fileName = fileName;
         this.versionList = versionList;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
