+++ src/main/java/de/hhbk/jahresprojekt/model/builder/DocumentBuilder.java
@@ -6,7 +6,7 @@ import de.hhbk.jahresprojekt.model.File;
 import java.util.List;
 
 public final class DocumentBuilder {
-    private int id;
+    private Long id;
     private String fileName;
     private List<File> versionList;
 
@@ -17,7 +17,7 @@ public final class DocumentBuilder {
         return new DocumentBuilder();
     }
 
-    public DocumentBuilder withId(int id) {
+    public DocumentBuilder withId(Long id) {
         this.id = id;
         return this;
     }
